Remove wrong versions
The app has version number 1.0.1 but you switched to the version number of the erpnext container instead of keeping the app versions like 1.0.2.
This causes the Software Center to pull the erpnext image http://ghcr.io/geniusdynamics/erpnext:15.69.2 instead of the app image. The erpnext image doesn't include the NS8 specific files like ui etc so it causes an error when installing.
The version of the erpnext container needs to be raised in the build-images.sh of the app.